app/requirements.txt
requests
colorama
Flask
python-dotenv
pillow
python-telegram-bot